modules/locations_service/requirements.txt
GeoAlchemy2==0.8.4
grpcio==1.67.0
grpcio-tools==1.67.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.0.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.9.9
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==0.16.1
kafka-python==2.2.11